requirements.txt
django.tailwindcss==3.2.7
django==4.2.5
ckeditor==6.4.2

#Bases de datos
mysqlclient==2.2.7

#forms
django-formtools
